start ai chat onboarding
